backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
requests==2.32.3
python-dotenv==1.2.2
